+++ src_clojure/clojure_sublimed/middleware.clj
@@ -1,6 +1,5 @@
 (ns clojure-sublimed.middleware
   (:require
-    [clojure.pprint :as pprint]
     [clojure.string :as str]
     [clojure-sublimed.core :as core]
     [nrepl.middleware :as middleware]
@@ -127,7 +126,3 @@
    :expects #{"eval" "clone"}
    :handles {"clone-and-eval"
              {:doc "Clones current session, evals given code in the cloned session"}}})
-
-(defn pprint [value writer opts]
-  (binding [pprint/*print-right-margin* 120]
-    (pprint/pprint value writer)))